+++ .vscode/launch.json
@@ -1,0 +1,28 @@
+{
+    // Use IntelliSense to learn about possible attributes.
+    // Hover to view descriptions of existing attributes.
+    // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+    "version": "0.2.0",
+    "configurations": [
+        {
+            "name": "Python: Current File",
+            "type": "python",
+            "request": "launch",
+            "program": "${file}",
+            "console": "integratedTerminal",
+            "args": [
+                "run",
+                "--port",
+                "5000"
+            ],
+            "env": {
+                "FLASK_APP": "app.app",
+                "REDIS_HOST":"127.0.0.1",
+                "FLASK_PORT": "8080"
+            },
+            "pythonPath": "${config:python.pythonPath}",
+            "cwd": "${workspaceFolder}",
+            "module": "flask"
+        }
+    ]
+}
