+++ app/templates/index.html
@@ -1,0 +1,7 @@
+{% block content %}
+<body>
+    <h1>{{ message }}</h1>
+
+   <img src="{{picture}}" alt="Smiley face" height="300" width="350">
+</body>
+{% endblock %}
