--- dcos/app.json
@@ -1,36 +1,0 @@
-{
-  "volumes": null,
-  "id": "/phippy/app",
-  "instances": 2,
-  "cpus": 1,
-  "mem": 128,
-  "disk": 0,
-  "gpus": 0,
-  "container": {
-    "type" : "DOCKER",
-    "docker": {
-      "image": "trainersontheweb/phippy:1.0",
-      "forcePullImage": false,
-      "privileged": false,
-      "network": "BRIDGE",
-      "portMappings": [
-        {
-          "name": "http",
-          "protocol": "tcp",
-          "containerPort": 80,
-          "hostPort": 0,
-          "servicePort": 1234,
-          "labels": {
-            "VIP_0": "app:80"
-          }
-        }
-      ]
-    }
-  },
-  "labels": {
-    "HAPROXY_GROUP": "external"
-  },
-  "env": {
-    "ETCD_HOST" : "db.marathon.l4lb.thisdcos.directory"
-  }
-}
