--- dcos/db.json
@@ -1,30 +1,0 @@
-{
-  "volumes": null,
-  "id": "/phippy/db",
-  "instances": 1,
-  "cpus": 1,
-  "mem": 128,
-  "disk": 0,
-  "gpus": 0,
-  "container": {
-    "type" : "DOCKER",
-    "docker": {
-      "image": "trainersontheweb/etcd:3.3.4",
-      "forcePullImage": false,
-      "privileged": false,
-      "network": "BRIDGE",
-      "portMappings": [
-        {
-          "name": "http",
-          "protocol": "tcp",
-          "containerPort": 4001,
-          "hostPort": 0,
-          "servicePort": 4001,
-          "labels": {
-            "VIP_0": "db:4001"
-          }
-        }
-      ]
-    }
-  }
-}
