+++ docker-compose.yml
@@ -1,0 +1,28 @@
+version: "3.7"
+
+services:
+  web:
+    image: hlesey/phippy:1.0
+    environment:
+      REDIS_HOST: redis
+      REDIS_PORT: 6379
+      FLASK_ENV: development
+    ports:
+      - "5000:5000"
+    networks:
+      - phippy
+    depends_on:
+      - redis
+  redis:
+    image: hlesey/redis:4
+    networks:
+      - phippy
+  nginx:
+    image: nginx
+    ports:
+      - "80:80"
+    networks:
+      - phippy
+networks:
+  phippy:
+    name: phippy
