--- docker/docker_build.sh
@@ -1,4 +1,0 @@
-#!/usr/bin/env bash
-
-VER=1.0
-docker build -t local/phippy:${VER} .
