--- docker/docker_cleanup.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-
-docker stop etcd phippy
-docker rm etcd phippy
-docker network rm phippy
