--- docker/requirements.txt
@@ -1,2 +1,0 @@
-flask
-python-etcd
