--- docker/templates/index.html
@@ -1,7 +1,0 @@
-{% block content %}
-<body>
-    <h1>{{ mesaj }}</h1>
-
-   <img src="{{poza}}" alt="Smiley face" height="300" width="350">
-</body>
-{% endblock %}
