+++ kubernetes/app.yaml
@@ -20,16 +20,18 @@ spec:
     spec:
       containers:
       - name: phippy
-        image: qedzone/phippy:1.0
+        image: hlesey/phippy:1.0
         env:
-          - name: ETCD_PORT
-            value: "4001"
+          - name: Reload
+            value: true
+          - name: REDIS_PORT
+            value: "6379"
         resources:
           requests:
             cpu: "100m"
             memory: "100Mi"
         ports:
-        - containerPort: 80
+        - containerPort: 5000
           protocol: TCP
           name: http
 ---
@@ -43,7 +45,7 @@ spec:
   ports:
   - name: http
     port: 80
-    targetPort: 80
+    targetPort: 5000
     protocol: TCP
 ---
 apiVersion: networking.k8s.io/v1beta1
